defacement/requirements.txt
pymysql
phpserialize
python-dotenv
requests